FINAL COMMIT before the semester submission
dev_mode.xml is for the checkbox ui, search_box.xml for the rest of the ui
 - both above xml files are included in activity_maps.xml

added information check for set current location button
 - the pngs are for the two new icons in the ui

there is also a new map overlay png - still needs work, but it's not bad for the time it took to change that (the last two weeks for me, ugh)

note the color change for the search bar hint text

set current location sort of works: it can set the search entry as the user location, but prolonged use of that function with the app currently causes unintended bugs. Also the routing has yet to work with this, but outside of that everything else should work...

...save for maybe the location and location updates. At least it routes! :)

And I think that's it. the apk is tucked away in BeachNav/app/build/output/apk (not really sure if that's the one though, you could try downloading this whole thing and try it anyhow)

...and i think that's most of the things i wanted to talk about.

- and that's that. - Austin T